pyproject.toml
[tool.poetry]
name = "ultibotinversiones"
version = "0.1.0"
description = "Plataforma de trading personal avanzada y de alto rendimiento."
authors = ["BMad Agent <bmad@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
pydantic = "^2.7.1"
pydantic-settings = "^2.3.0" # Compatible con Pydantic 2.x
uvicorn = {extras = ["standard"], version = "^0.29.0"}
langchain-core = ">=0.3.62,<0.4.0"
langchain-google = "*"
langchain-community = "*"
supabase = "^2.4.0"
httpx = "^0.27.0"
websockets = "^12.0"
asyncpg = "^0.30.0"
cryptography = "^45.0.3"
psycopg = {extras = ["binary", "pool"], version = "^3.2.9"}
langchain-google-genai = "^2.1.5"
pyparsing = "^3.2.3"
pyqtgraph = "^0.13.3"
qdarkstyle = "^3.1"
qasync = "*"
mplfinance = "*"
tenacity = "^8.3.0"
asgi-correlation-id = "^4.3.1"
python-binance = "*"
aiohttp = "*"
pyside6 = {version = ">=6.5.0,<6.10.0", python = ">=3.11,<3.14"}
psutil = "^7.0.0"
injector = "^0.21.0"


[tool.poetry.group.dev.dependencies]
ruff = "*"
pytest = "*"
pytest-asyncio = "*"
pytest-cov = "*"
pre-commit = "*"
pytest-dotenv = "^0.5.2"

[tool.poetry.scripts]
test = "pytest:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "UP", # pyupgrade
    "D",  # pydocstyle
    "B",  # flake8-bugbear
]
# ignore = [] # Specific rules to ignore

[tool.ruff.format]
# Ruff's formatter is Black-compatible by default.
# Options can be added here if specific deviations from Black are needed.
# quote-style = "double"
# indent-style = "space"
# skip-magic-trailing-comma = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "*Tests"
python_functions = "test_*"
dotenv_files = ".env.test"
addopts = "-p no:warnings"
